ct: move l1 objects into l1 package
