+++ src/v/cloud_topics/level_one/object.h
