--- src/v/cloud_topics/objects/tests/BUILD
@@ -1,17 +1,0 @@
-load("//bazel:test.bzl", "redpanda_cc_gtest")
-
-redpanda_cc_gtest(
-    name = "level_one_object_test",
-    timeout = "short",
-    srcs = [
-        "level_one_object_test.cc",
-    ],
-    deps = [
-        "//src/v/bytes:iostream",
-        "//src/v/cloud_topics/objects:level_one_object",
-        "//src/v/model/tests:random",
-        "//src/v/test_utils:gtest",
-        "@googletest//:gtest",
-        "@seastar",
-    ],
-)
